[Feat] ResourceDegenerator 구현 및 Hunger수정
