+++ Assets/KSH/Scripts/Interface/IResource.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using System;
 
 namespace KSH
 {
