+++ Assets/KSH/Scripts/Interface/ITickableResource.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-
 namespace KSH
 {
     public interface ITickableResource
