+++ Assets/KSH/Scripts/Resource.cs
@@ -1,7 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Events;
+using System;
 using KSH;
 
 namespace KSH
@@ -11,9 +11,8 @@ namespace KSH
         [Header("Resource Amount")]
         [Range(0, 100)][SerializeField] private float max;//최대 자원 양
         [Range(0, 100)] [SerializeField] private float cur; //현재 자원 양
-
-		[Header("Event")]
-		public UnityEvent OnResourceChanged; //이벤트 
+        
+        public event Action<Resource> OnResourceChanged; //코드용 이벤트
         
         public float Max => max;
         public float Cur => cur;
@@ -24,7 +23,7 @@ namespace KSH
             cur = Mathf.Clamp(cur - amount, 0, max); // 최소, 최대 제한 값 설정
             if (Mathf.Abs(beforeCur - cur) > 0.01f) 
             {
-                OnResourceChanged?.Invoke(); // 현재 자원양과 전의 양의 차이가 생기면 이벤트 발생
+                OnResourceChanged?.Invoke(this); // 현재 자원양과 전의 양의 차이가 생기면 이벤트 발생
             }
         }
         public void Increase(float amount) //자원 증가
@@ -33,7 +32,7 @@ namespace KSH
             cur = Mathf.Clamp(cur + amount, 0, max); // 최소, 최대 제한 값 설정
             if (Mathf.Abs(cur - beforeCur) > 0.01f) 
             {
-                OnResourceChanged?.Invoke(); // 현재 자원양과 전의 양의 차이가 생기면 이벤트 발생
+                OnResourceChanged?.Invoke(this); // 현재 자원양과 전의 양의 차이가 생기면 이벤트 발생
             }
         }
     }
