+++ Assets/KSH/Scripts/ResourceDegenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0701b88209cc0c45a292d645e7dc6ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
